requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-crispy-forms==2.3
django-environ==0.11.2
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
crispy-bootstrap5
crispy-bootstrap5
channels
